fix(acr): run QA through package manager entrypoint
